run: Do not hold the lock while creating the database
Until now, the global db cache lock was held until the db of
the playground was created. To reduce the lock contention when
we get large amount of simultaneous requests, only take the lock
when reading or writing to the db cache.

This change requires to introduce a wait mechanism, in the case of
two simultaneous requests on the same database, when the db has not
been created yet. In this case, the second request will poll the
cache every 5ms to check if the db is ready. There's no specific
reason for choosing 5ms, so we might want to adjust this in the
future.